vision-mamba-control/requirements.txt
# Vision Mamba Control - Requirements

# Deep Learning
torch>=2.0.0
einops>=0.7.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# GUI
# tkinter (built-in with Python)

# Utilities
numpy>=1.24.0
loguru>=0.7.0

# Optional - for training (not needed for demo)
# torchvision>=0.15.0
# tqdm>=4.65.0
# matplotlib>=3.7.0
